Shared DEGs of all paired comparisons in all samples expression dataframe of RNA-Seq.

Shared DEGs of all paired comparisons in all samples expression dataframe of RNA-Seq.

Usage

data(gene_expression2)

Format

Dataframe: Shared DEGs of all paired comparisons in all samples expression dataframe of RNA-Seq. (1st-col: Genes, 2nd-col~: Samples).

References

https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CorPlot/

Author

benben-miao

Examples

# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset
data(gene_expression2)

# 3. View gene_expression2
gene_expression2
#>          Genes   CT_1    CT_2    CT_3  LT20_1  LT20_2  LT20_3  LT15_1  LT15_2
#> 1        ACAA2  24.50   39.83   55.38  114.11  159.32   96.88  169.56  464.84
#> 2         ACAN  14.97   18.71   10.30   71.23  142.67  213.54  253.15  320.80
#> 3         ADH1   1.54    1.56    2.04   14.95   13.60   15.87   12.80   17.74
#> 4         AHSG   0.00 1911.99    0.00    0.00    0.00    0.00    0.00    0.00
#> 5        ALDH2   2.07    2.86    2.54    0.85    0.49    0.47    0.42    0.13
#> 6        AP1S3   6.62   14.59    9.30   24.90   33.94   23.19   24.00   36.08
#> 7        APOA1  64.50   38.33   45.89   97.49   87.95  131.96  176.52  157.89
#> 8         APOB  35.72    0.00    0.00    0.00    0.00    0.00    0.00    0.00
#> 9        APOC2   2.62    3.95    6.04 1479.08 2025.74 1924.97 2291.30 2505.23
#> 10       APOEB   1.14    0.21    1.40    1.97   14.37   13.91   11.71    4.60
#> 11    ARHGEF28  24.16   24.72   29.19   12.38   13.04   10.96    6.42    5.05
#> 12      ARRDC2  22.96   26.98   38.65    7.92    9.59    3.10    4.08    3.91
#> 13       BLVRB  12.44   16.36   14.78    4.54    5.52    5.43    4.49    2.35
#> 14        BTG3  13.07   14.47    9.06   32.95   28.50   35.01   38.48   71.63
#> 15          C3  38.14   44.59   38.81   61.57  112.18   89.48   79.86  151.22
#> 16          C5   0.00    0.00    0.00    0.00  129.11  127.14  122.14   90.05
#> 17          C7   1.36    2.42    3.00    3.34    5.62   18.17   48.29   30.58
#> 18          C9   2.80    2.35    6.10    9.79   17.00    9.61    9.01    8.44
#> 19       CCNG2  10.70   25.13   20.20    7.03   11.31    7.96    6.64    7.39
#> 20        CDO1  23.13   62.92   36.00  133.92  157.69  256.44  429.78  431.33
#> 21       CERS2   7.44    7.33    7.86   24.54   15.72   20.16   23.80   26.51
#> 22         CFB   0.00    0.00    0.00   52.69   65.14   12.54   22.84   14.98
#> 23         CFH   0.00    0.00    0.00   13.69    0.00    0.00    2.03    0.00
#> 24        CISH  23.48   12.68   21.64    9.47    3.30    4.83    4.64    4.44
#> 25      CLEC4E  31.36   37.16   46.40   83.45   82.60  116.26  108.56  137.31
#> 26      CLEC4M  27.21   31.86   35.85   47.28   54.16  109.35   76.82  119.81
#> 27          CP   0.00    0.00    0.00    0.00    0.00   67.17   53.75   28.23
#> 28         CYC   1.97    0.65    1.96    3.92    3.49   10.60   18.68   48.73
#> 29     CYP24A1   0.76    0.27    0.00    3.30    6.09    6.87    3.00  355.44
#> 30      CYP2J2   0.00    0.00    0.00    0.00    0.00  124.21   97.65  127.76
#> 31      CYP2K1   0.00    0.00    9.46    0.00    0.00    0.00    0.00    0.00
#> 32       DECR2  12.07   15.51   21.24    6.08    7.44    6.49    6.27    4.82
#> 33      DHCR24   9.09   14.86   10.59   18.05   48.64   31.89   37.95   42.83
#> 34       DHCR7   5.73    5.85    5.70    9.31   18.11   22.00   22.01   22.39
#> 35        DIO3  57.97   21.74   51.13  168.89  115.64  121.29   67.29  203.51
#> 36        DLC1   2.75    3.41    3.55    7.50    7.96   12.30    9.27   12.12
#> 37        EEF2   0.00    0.00    0.09    0.00    0.00 1060.87    0.00   37.82
#> 38       EIF4E   2.17    1.19    3.29   13.34    5.29   10.84   18.60   48.02
#> 39         ELL   5.87    5.25    5.75   16.67   11.19   20.47   16.68   32.75
#> 40      ERRFI1   1.78    6.74    7.02   56.54   34.02   77.81   70.98  121.53
#> 41          F2   2.27    7.98    8.74   31.91   33.33   36.08  111.50  113.33
#> 42        FAAH 164.88   67.27  127.87   24.68    9.03   16.24    4.19    2.86
#> 43        FBP1  88.80   61.88   48.96  168.77  146.22  131.51  372.23   86.25
#> 44        FBP2   4.54   10.11    7.53   17.07   18.45   13.82   67.46   10.03
#> 45        FDPS   9.96    8.54    8.79   15.83   25.53   23.29   26.46   32.14
#> 46       FETUB   0.00    0.00    0.00    0.00    0.00   24.34   13.05    0.00
#> 47         FGA   0.00    0.00    0.07  307.83    0.00    0.00    0.00  101.38
#> 48         FGB   0.48    1.70    0.74   17.61    6.39   12.85   23.39   20.45
#> 49        FHL1   5.75    8.37    7.84    3.74    3.57    2.92    3.22    2.33
#> 50       FKBP5   0.00    0.46    0.32    4.04    2.07    5.51    6.61    9.44
#> 51         FN1   0.00    0.00   22.50    0.00    0.00    0.00    0.00    0.00
#> 52       FOSL2   0.49    0.42    0.50    3.19    1.42    2.40    3.41    9.47
#> 53        GPD1  83.21   62.27   57.36  108.42  345.17  139.29  201.86  304.92
#> 54       GRINA   8.53    8.83   10.08   18.91   29.46   34.40   20.94   44.00
#> 55        HAMP   7.78   13.47   82.53  190.21  393.82  297.98  498.86   77.94
#> 56        HEPH   0.00    0.00    0.14    0.00    0.00  234.50  230.72  163.77
#> 57      HIGD1A   3.21    3.23    2.33    9.65    4.99   17.90   11.59   21.49
#> 58         HLF   1.72    1.79    1.11    7.41    3.13    7.26    9.04    3.96
#> 59       HMGCR   0.00    0.00   51.65    0.00    0.00    0.00    0.00    0.00
#> 60      HMGCS1   3.49    1.27    4.49    6.68   30.43   11.01   10.55   25.40
#> 61        HMOX 226.67  204.06  166.60  108.74   48.59   97.45   58.39   78.18
#> 62         HPX   0.00    0.00    0.00    0.00 9651.15    0.00 8446.26 8481.58
#> 63    HS3ST3B1   4.86    5.10    6.65   14.92    8.54   15.87   12.40   16.02
#> 64     HSD17B7   1.47    1.98    0.88    3.60    7.25    5.12    6.79    7.70
#> 65       HSP30   0.95    1.55    1.35    0.00    0.11    0.00    0.00    0.00
#> 66        IDI1   6.14    4.62    6.06    9.21   34.69   11.46   13.36   24.57
#> 67        IDO2  27.14   32.27   33.07   16.19   10.34   12.29   11.23    9.44
#> 68       IL1R1  15.26   16.76   20.95   37.52   30.27   50.57   70.69  100.78
#> 69        IL6R  10.16   18.90   16.35   28.08   23.56   54.78   42.13   62.81
#> 70      INSIG1   2.10    1.33    1.56    2.81    4.30    7.07    9.59   15.46
#> 71       ITIH3  16.09    0.00   37.11    0.00    0.00    0.00    0.00    0.00
#> 72       KLF12   2.67    2.59    1.89    7.65    4.61   12.22   14.14   12.14
#> 73       KRT18 165.53    0.02    0.00    0.09    0.08    0.00    0.00    0.00
#> 74      L2HGDH   0.86    1.07    0.47   10.56   24.52   30.27   73.10   41.90
#> 75       LECT2   7.68    8.98    7.87   29.97   23.54   19.79   41.79   23.36
#> 76        LRR1   3.99    3.81    4.42    2.24    1.04    0.49    0.46    0.66
#> 77       MAPK6   9.51    7.95    7.68   34.49   25.80   46.91   60.90   95.99
#> 78         MDH  53.18   80.20   65.10  145.18  200.30  220.68  332.68  274.00
#> 79    MID1IP1L  15.70   13.61   27.96    7.84    3.27    3.56    1.17    3.80
#> 80       MKNK2   6.62    5.33    6.33   13.34   12.37   11.36   19.26   33.84
#> 81     MT-ATP6   0.00    0.00 8156.59    3.42    0.00    0.00    0.00    0.00
#> 82      MT-CO1  25.31   30.96 1899.15   20.95   29.74   21.02   18.55   16.75
#> 83      MT-ND2   0.49    0.34 2054.63    0.37    0.34    0.25    0.45    0.14
#> 84      MT-ND4   0.00    0.00 1766.51    0.00    0.00    0.00    0.00    0.00
#> 85      MT-ND6   0.00    0.00   91.69    0.00    0.00    0.00    0.00    0.00
#> 86       NFIL3  11.91   10.62   12.54   25.88   28.23   28.37   28.69   40.49
#> 87        NOCT   5.38    4.00    3.08   17.56   14.62   26.98   20.26   25.41
#> 88        NQO1  27.81   25.53   32.60   15.53   11.04   10.91   13.82   11.97
#> 89       NUDT7   1.26    2.19    1.82    0.79    0.78    0.67    0.56    0.57
#> 90    P33MONOX   5.68   18.29    9.32    1.12    2.25    1.32    3.45    0.58
#> 91      PAIP2B   4.25    8.82    7.82   18.87   16.21   27.52   23.27   27.97
#> 92       PDIA3   6.03   11.91   10.81   14.73   44.90   15.73   18.39   18.02
#> 93        PDK2  56.43   52.68   68.44  173.60  131.41  153.44  150.22  136.72
#> 94      PDLIM2   7.74   12.63   12.18    2.42    5.56    4.62    4.90    2.41
#> 95        PIM3   4.22    5.58    4.46   12.75   13.91   16.59   32.96   51.26
#> 96       PLIN2  21.51   16.54   14.57   31.51   57.29  104.49  118.31  150.54
#> 97         PNP  85.70   91.05   42.92  202.07  193.91  601.36  733.04 1020.55
#> 98       PPARA   1.70    2.78    2.03    3.47   12.36   22.82   23.27   18.83
#> 99      PPDPFA  86.78  145.93   92.48  217.31  284.30  244.63  369.37  408.03
#> 100      PPM1K   2.22    0.98    0.65    1.75   16.28    9.13   20.59    8.73
#> 101   PPP1CC-A   1.08    2.40    1.21    5.10    6.30   10.26   17.49   20.43
#> 102    PPP1R3B  34.32   64.52   18.30  184.01  136.24  175.78  121.00  343.29
#> 103      PPTC7   0.39    0.48    0.50    3.68    3.73   17.80    9.69    3.77
#> 104   PRELID3B   3.59    2.25    2.48   16.53   19.39   35.82   36.12   98.04
#> 105      PRODH   0.15    0.17    0.20    0.96    0.58    5.18   15.49   35.66
#> 106       PSPH   2.34    1.56    2.66    3.56    5.81    3.91   10.73   11.15
#> 107      RASD2   0.48    0.24    0.12    3.59    3.74    8.09    2.51    6.45
#> 108       RBMX  13.64   14.73   13.16    5.21    6.14    7.00    4.55    3.96
#> 109      RBP4A 204.93  257.76  408.01   99.21  178.89   91.27  142.39   55.52
#> 110     RHBG-A  21.57   13.77   31.06   11.97    7.78    6.14    3.17    3.43
#> 111      RHOAD   3.40    3.72    3.73   11.18    6.54   11.38    9.96   10.59
#> 112       RND3   8.24    8.27    9.73   10.99   25.46   26.41   33.99   30.30
#> 113     RNF223   0.08    0.28    0.21    1.89    4.04    2.93    1.96    3.54
#> 114      RRBP1  11.50   13.82   17.93   19.05   43.28   26.91   31.11   59.23
#> 115      RXRGB  43.66   27.43   37.62   13.00   11.92    9.17   11.17    9.56
#> 116       SAT1  11.56   12.57    8.49   33.31   14.73   36.37   41.34   48.72
#> 117 SERPINA3-3   0.00    0.00    0.00   22.69   12.99    7.21   33.92   10.06
#> 118 SERPINA3-8  25.72   29.82   47.98   61.57   61.35   84.34  114.27  111.11
#> 119   SERPING1   0.00    0.00    0.00    0.00   80.00    0.00  125.16   60.04
#> 120       SFPQ  16.34   13.54   22.15   37.51   31.26   30.94   60.73   59.24
#> 121    SLC16A1   0.00    0.84    0.57    6.43    3.74   26.14   42.09   17.62
#> 122   SLC20A1B   5.89    6.01    7.72    9.84   13.44   33.19   25.94   59.93
#> 123    SLC22A2  35.76   27.33   13.35   11.87    4.01    6.47    4.86    3.06
#> 124  SLC25A25A   0.61    2.00    0.24    9.24    8.43   49.29   54.00   96.91
#> 125    SLC2A12  10.66   11.44   12.25    2.95    6.47    5.74    4.18    2.35
#> 126     SLC3A2   3.14    3.59    3.87    7.93    6.87    8.03    9.79   18.23
#> 127      SOAT1  12.37   15.43   19.51   79.14   46.70   80.64   73.78  144.14
#> 128      SOCS3   1.42    2.29    1.28    4.41    4.06    3.92    3.12    4.51
#> 129       SSR1   0.00    0.00    0.00    0.00   26.46   19.42   20.87   10.95
#> 130        TAT   0.76    0.50    0.47    2.66    5.77   11.38    9.14    9.12
#> 131      TCF21   5.79    5.62    8.18    3.44    3.16    2.30    1.74    2.01
#> 132         TF   0.00    0.00    0.00 2252.09    0.00 1005.29    0.00 1314.18
#> 133      TOR3A   1.20    1.35    0.77    2.59    6.68   15.60    5.52   11.13
#> 134       UGT3   0.76    2.47    2.39    6.71    6.61    6.28    5.64    6.25
#> 135       UPP2  62.04   30.85   22.86  130.71  119.35   70.28  107.05   89.01
#> 136       UROS   5.52   10.21    7.86    2.10    2.71    3.03    0.71    0.79
#> 137       XBP1  67.53   40.89   80.11  112.77  149.74  204.12  189.54  173.80
#> 138     YME1L1   9.12    8.29   10.70   17.32   15.37   22.24   24.53   40.04
#> 139      YPEL3  17.44   29.03   27.49    6.63    7.95    4.24    2.89    2.44
#>      LT15_3  LT12_1  LT12_2  LT12_3 LT12_6_1 LT12_6_2 LT12_6_3
#> 1    182.66  116.08  497.29  464.48   471.43   693.62   229.77
#> 2    104.15  174.02  305.81  469.48  1291.90   991.90   966.77
#> 3      6.06   10.97   10.71   30.95     9.84    10.91     7.28
#> 4      0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 5      0.26    0.00    0.28    0.11     0.37     0.15     0.11
#> 6     27.40   24.06   38.74   34.54    62.72    41.36    28.75
#> 7    150.87  168.06  116.12   95.61   364.08   163.94   278.73
#> 8      0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 9      9.69 3265.83 2678.45 2806.98   995.99     3.06  2403.92
#> 10     4.70    2.15   37.11    5.85    33.07     3.11     2.78
#> 11     5.60   10.73    3.04    4.77     1.52     1.36     4.18
#> 12     4.81    4.33    3.47    3.94     2.65     3.19     2.84
#> 13     7.64    7.21    3.44    3.25     2.70     2.25     3.05
#> 14    51.61  214.30  182.88  141.60   218.96   233.89   263.17
#> 15   100.95   81.91  131.64  167.63   266.65   237.92   138.29
#> 16   110.96  142.83   80.80    0.00     0.00    80.73    77.76
#> 17    96.20   10.74   14.04   21.74   606.46    32.79    29.88
#> 18     9.54   15.21   15.03   13.50     6.99    12.21     6.62
#> 19    10.38    6.72    3.86    7.34     3.10     2.70     2.98
#> 20   193.86  182.78  335.23  387.30   998.72   571.06   694.93
#> 21    19.79   33.66   24.41   28.06    40.22    32.93    40.69
#> 22     0.00   19.19    5.39    6.23     5.69    21.69     0.83
#> 23    24.41    0.13   13.16   26.10    28.65    16.51     5.97
#> 24    10.03    6.97    1.44    4.80     0.29     0.65     3.73
#> 25    93.62   88.49  106.60  129.20   163.32   187.30   129.81
#> 26    65.44   57.43   70.83  112.74   140.70   141.62   118.59
#> 27    44.38   70.26   25.70    0.00     0.00     0.00    84.71
#> 28    15.65   41.75   76.07   36.24   214.22   242.84   174.13
#> 29     1.90  421.67   12.88  491.46     1.30     5.55   265.26
#> 30   132.08   89.19   98.59    0.00     0.00     0.00    91.88
#> 31     0.00    0.01    0.00    0.00     0.00     0.00     0.00
#> 32     8.09    4.39    6.14    4.50     4.05     5.32     3.72
#> 33    18.72   26.58   42.72   33.34    33.62    26.40    28.55
#> 34    12.23   17.90   13.36   10.68    21.26     8.10    13.82
#> 35   129.34  287.82  204.22  233.76   309.44   249.64   427.45
#> 36     8.47   11.35   14.76   14.86    15.72    17.26    17.43
#> 37    69.62  659.33   77.25    0.00     0.00     0.03   626.92
#> 38    18.84   23.74   38.91   46.91   159.22   163.60   161.56
#> 39    14.02   29.65   21.86   30.89    18.04    24.80    28.82
#> 40    63.23   93.40   88.13  127.68    85.17   157.47   144.66
#> 41    33.09   82.80   96.58   50.05   109.73    80.17    98.46
#> 42    10.12    7.78   10.39    3.08     6.17     1.29     1.75
#> 43   301.95  254.91  277.23  231.60   691.82   449.71   284.45
#> 44    36.18   41.97   56.82   38.45   161.39   132.34    79.71
#> 45    15.45   21.58   39.04   17.56    19.43    16.86    17.87
#> 46     0.00   14.99   16.43   36.53    26.05    13.40     6.20
#> 47   104.37    0.00   88.60    0.00     0.00     0.23    99.65
#> 48    11.49   41.83   46.80   42.01    58.47    39.46    51.17
#> 49     1.79    3.70    2.60    2.27     1.54     1.71     2.61
#> 50     4.18    8.50   10.97    6.87    15.44    21.05    15.56
#> 51     0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 52     7.92   58.44   56.24   45.64   238.59   167.53   239.07
#> 53   169.81  173.53  385.57  159.97   204.50    89.88   120.77
#> 54    38.98   44.57   29.90   44.62    40.50    32.11    32.94
#> 55   201.98 1849.52   76.20  637.73   560.96   467.06    73.16
#> 56   203.27  299.75  176.45    0.00     0.00     0.00   159.42
#> 57    14.44   43.94   53.59   18.21   162.00   205.78   144.63
#> 58     8.40   10.75    6.83    6.20     4.30     2.65     4.16
#> 59     0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 60     7.10    8.63   17.12   18.04    19.55    13.14     9.87
#> 61    49.52   54.23   65.29   58.87    30.40    17.50    41.79
#> 62  6142.08 6400.80 6676.96    0.00     0.00     0.00  7357.38
#> 63     8.95    9.30   11.73   18.16    37.33    56.39    30.74
#> 64     3.12    2.38   10.35   10.06     4.58     7.32     5.32
#> 65     0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 66     8.61    9.37   21.64   14.79    15.45    12.73     9.72
#> 67    18.20   13.76    7.05    6.46     4.88     5.84    12.94
#> 68    56.56   41.50   91.08   66.61   105.81    67.13    74.22
#> 69    23.52   50.14   60.56   80.53    81.77   102.31    95.44
#> 70    10.37   23.65   33.86   19.15   130.20   166.01   130.09
#> 71     0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 72     6.04   10.88    9.87   11.30    18.74    12.68    18.19
#> 73     0.00    0.00    0.04    0.00     0.00     0.05     0.03
#> 74    57.37   57.96   72.51   26.20    46.17    65.90    60.16
#> 75    15.87   72.84   47.15   61.46    87.96   124.25    14.35
#> 76     0.98    1.40    0.67    0.99     0.28     0.29     0.42
#> 77    73.19  129.77  114.50  104.64   146.49   186.07   165.92
#> 78   243.35  219.34  354.22  228.59   413.66   361.52   326.84
#> 79     1.74    4.34    0.37    2.32     0.76     0.20     4.62
#> 80    18.84   37.46   31.46   32.56    45.56    46.89    42.67
#> 81     0.00    0.00    2.59    0.00     0.00     0.00     6.60
#> 82    22.88   34.56   21.82   22.73    31.83    40.78    22.58
#> 83     0.17    0.10    0.00    0.21     0.12     0.00     0.00
#> 84     0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 85     0.00    0.00    0.00    0.00     0.00     0.00     0.00
#> 86    23.54   70.55   66.53   45.30    45.03    68.50    60.23
#> 87    24.74   84.77   58.77   22.73    35.46    64.57    52.26
#> 88    15.35   14.51   13.31   12.50     9.53     8.92    11.16
#> 89     0.46    0.49    0.26    0.15     0.06     0.19     0.05
#> 90     2.46    3.88    1.60    1.06     3.46     3.72     2.97
#> 91    36.81   22.12   22.15   26.09    13.12    13.67    20.54
#> 92    22.82   15.96   26.83   26.07    33.92    54.45    43.74
#> 93    85.82  191.83  217.73  131.16   179.39    76.23   186.22
#> 94     2.28    3.68    1.69    2.08     0.28     0.59     0.84
#> 95    26.56   64.29   78.77   91.87    96.73    86.93    77.18
#> 96    74.17  290.93  419.72  132.97   540.67   232.39   228.76
#> 97   890.82 1575.95 1198.74 1408.24  1175.85  2357.23  1258.27
#> 98    11.35   13.43   57.39    7.85   171.78   148.18    74.50
#> 99   383.80  245.51  465.59  484.83   237.96   236.50   257.21
#> 100    6.74    2.49   21.55    5.08    39.17     9.03     5.75
#> 101    7.86   26.69   19.36   25.87     6.99    13.08     8.62
#> 102  200.69  506.26  190.43  288.98    60.34   132.70   228.27
#> 103   12.58    5.35    1.50    4.07     2.01     1.56     0.68
#> 104   37.35  132.31  190.23   65.66   279.82   332.08   264.18
#> 105   10.03   20.23   35.74   37.68   416.51   262.57   219.47
#> 106    8.40    3.02   17.95    7.05    37.24    18.54    10.32
#> 107    2.44    2.82    5.87   27.05    22.99    31.24     4.28
#> 108    3.17    2.16    3.23    4.05     4.55     2.41     2.97
#> 109   91.47  134.51  133.66   61.42    36.71    28.34    38.64
#> 110    4.51    6.29    4.62    3.30     1.47     2.78     1.34
#> 111    6.40    5.70    9.91    8.05     9.05     9.27     8.01
#> 112   20.58   32.17   53.71   22.12    64.52    49.27    46.64
#> 113    0.81    1.21    1.63    3.86     3.74     9.43     1.82
#> 114   34.86   18.92   47.27   38.44    44.65    69.21    67.59
#> 115   18.85   14.01   11.54   11.63     6.55     4.71     9.13
#> 116  111.05  211.18  239.67   90.39   351.13   248.68   265.93
#> 117  121.29   97.03   98.05  175.54     1.04    15.33     7.42
#> 118  104.81  191.13  203.26  135.33   458.27   337.34   267.87
#> 119    0.00    0.00   90.52    0.00     0.00     0.00   167.64
#> 120   39.01   60.93   64.89   58.53    83.57    82.34    91.65
#> 121   11.76   37.55   29.55   23.75   105.30    53.47    89.69
#> 122   21.43   43.27   78.51   36.51   201.82   236.09   202.91
#> 123    8.47    9.18    2.46    3.72     2.02     6.83     4.24
#> 124   41.88  144.65   98.98   93.93    96.27   131.40   116.82
#> 125    6.28    4.89    2.38    3.56     0.55     0.44     1.20
#> 126   11.29   16.29   26.46   13.43    65.30    55.32    61.76
#> 127  169.49   64.24  117.66   58.23    20.91    62.00    58.93
#> 128    6.12   17.00   14.23   11.56    14.72    15.31    12.27
#> 129   36.22    8.16   33.12    0.00    27.40    26.50     0.00
#> 130   11.19   16.10    4.89   11.53     6.90     9.36     6.22
#> 131    1.81    1.27    1.02    1.53     1.46     0.77     1.95
#> 132 1548.00 4853.35  967.87 1399.18     0.00  1016.56   718.36
#> 133    5.37   10.20   18.64    9.09    32.27    39.23    18.66
#> 134   29.04    3.58   10.61    5.35     5.71    19.50    22.82
#> 135  242.80  124.80  113.77  171.00   120.73    86.24   189.38
#> 136    3.39    1.15    0.56    0.65     0.56     0.88     0.42
#> 137  130.33  192.53  217.46  161.70   241.69   267.28   183.90
#> 138   22.77   27.85   34.74   30.47    54.87    52.23    60.78
#> 139    9.38    6.13    4.14    4.47     0.62     1.22     1.40