Shared DEGs of all paired comparisons in all groups expression dataframe of RNA-Seq.

Shared DEGs of all paired comparisons in all groups expression dataframe of RNA-Seq.

Usage

data(gene_expression3)

Format

Dataframe: Shared DEGs of all paired comparisons in all groups expression dataframe of RNA-Seq. (1st-col: Genes, 2nd-col~n-1-col: Groups, n-col: Pathways).

References

https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/CorPlot/

Author

benben-miao

Examples

# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example dataset
data(gene_expression3)

# 3. View gene_expression3
gene_expression3
#>          Genes           CT         LT20         LT15         LT12       LT12_6
#> 1        ACAA2 3.990333e+01 1.234367e+02  272.3533333 3.592833e+02 4.649400e+02
#> 2         ACAN 1.466000e+01 1.424800e+02  226.0333333 3.164367e+02 1.083523e+03
#> 3         ADH1 1.713333e+00 1.480667e+01   12.2000000 1.754333e+01 9.343333e+00
#> 4         AHSG 6.373300e+02 0.000000e+00    0.0000000 0.000000e+00 0.000000e+00
#> 5        ALDH2 2.490000e+00 6.033333e-01    0.2700000 1.300000e-01 2.100000e-01
#> 6        AP1S3 1.017000e+01 2.734333e+01   29.1600000 3.244667e+01 4.427667e+01
#> 7        APOA1 4.957333e+01 1.058000e+02  161.7600000 1.265967e+02 2.689167e+02
#> 8         APOB 1.190667e+01 0.000000e+00    0.0000000 0.000000e+00 0.000000e+00
#> 9        APOC2 4.203333e+00 1.809930e+03 1602.0733330 2.917087e+03 1.134323e+03
#> 10       APOEB 9.166667e-01 1.008333e+01    7.0033333 1.503667e+01 1.298667e+01
#> 11    ARHGEF28 2.602333e+01 1.212667e+01    5.6900000 6.180000e+00 2.353333e+00
#> 12      ARRDC2 2.953000e+01 6.870000e+00    4.2666667 3.913333e+00 2.893333e+00
#> 13       BLVRB 1.452667e+01 5.163333e+00    4.8266667 4.633333e+00 2.666667e+00
#> 14        BTG3 1.220000e+01 3.215333e+01   53.9066667 1.795933e+02 2.386733e+02
#> 15          C3 4.051333e+01 8.774333e+01  110.6766667 1.270600e+02 2.142867e+02
#> 16          C5 0.000000e+00 8.541667e+01  107.7166667 7.454333e+01 5.283000e+01
#> 17          C7 2.260000e+00 9.043333e+00   58.3566667 1.550667e+01 2.230433e+02
#> 18          C9 3.750000e+00 1.213333e+01    8.9966667 1.458000e+01 8.606667e+00
#> 19       CCNG2 1.867667e+01 8.766667e+00    8.1366667 5.973333e+00 2.926667e+00
#> 20        CDO1 4.068333e+01 1.826833e+02  351.6566667 3.017700e+02 7.549033e+02
#> 21       CERS2 7.543333e+00 2.014000e+01   23.3666667 2.871000e+01 3.794667e+01
#> 22         CFB 0.000000e+00 4.345667e+01   12.6066667 1.027000e+01 9.403333e+00
#> 23         CFH 0.000000e+00 4.563333e+00    8.8133333 1.313000e+01 1.704333e+01
#> 24        CISH 1.926667e+01 5.866667e+00    6.3700000 4.403333e+00 1.556667e+00
#> 25      CLEC4E 3.830667e+01 9.410333e+01  113.1633333 1.080967e+02 1.601433e+02
#> 26      CLEC4M 3.164000e+01 7.026333e+01   87.3566667 8.033333e+01 1.336367e+02
#> 27          CP 0.000000e+00 2.239000e+01   42.1200000 3.198667e+01 2.823667e+01
#> 28         CYC 1.526667e+00 6.003333e+00   27.6866667 5.135333e+01 2.103967e+02
#> 29     CYP24A1 3.433333e-01 5.420000e+00  120.1133333 3.086700e+02 9.070333e+01
#> 30      CYP2J2 0.000000e+00 4.140333e+01  119.1633333 6.259333e+01 3.062667e+01
#> 31      CYP2K1 3.153333e+00 0.000000e+00    0.0000000 3.333333e-03 0.000000e+00
#> 32       DECR2 1.627333e+01 6.670000e+00    6.3933333 5.010000e+00 4.363333e+00
#> 33      DHCR24 1.151333e+01 3.286000e+01   33.1666667 3.421333e+01 2.952333e+01
#> 34       DHCR7 5.760000e+00 1.647333e+01   18.8766667 1.398000e+01 1.439333e+01
#> 35        DIO3 4.361333e+01 1.352733e+02  133.3800000 2.419333e+02 3.288433e+02
#> 36        DLC1 3.236667e+00 9.253333e+00    9.9533333 1.365667e+01 1.680333e+01
#> 37        EEF2 3.000000e-02 3.536233e+02   35.8133333 2.455267e+02 2.089833e+02
#> 38       EIF4E 2.216667e+00 9.823333e+00   28.4866667 3.652000e+01 1.614600e+02
#> 39         ELL 5.623333e+00 1.611000e+01   21.1500000 2.746667e+01 2.388667e+01
#> 40      ERRFI1 5.180000e+00 5.612333e+01   85.2466667 1.030700e+02 1.291000e+02
#> 41          F2 6.330000e+00 3.377333e+01   85.9733333 7.647667e+01 9.612000e+01
#> 42        FAAH 1.200067e+02 1.665000e+01    5.7233333 7.083333e+00 3.070000e+00
#> 43        FBP1 6.654667e+01 1.488333e+02  253.4766667 2.545800e+02 4.753267e+02
#> 44        FBP2 7.393333e+00 1.644667e+01   37.8900000 4.574667e+01 1.244800e+02
#> 45        FDPS 9.096667e+00 2.155000e+01   24.6833333 2.606000e+01 1.805333e+01
#> 46       FETUB 0.000000e+00 8.113333e+00    4.3500000 2.265000e+01 1.521667e+01
#> 47         FGA 2.333333e-02 1.026100e+02   68.5833333 2.953333e+01 3.329333e+01
#> 48         FGB 9.733333e-01 1.228333e+01   18.4433333 4.354667e+01 4.970000e+01
#> 49        FHL1 7.320000e+00 3.410000e+00    2.4466667 2.856667e+00 1.953333e+00
#> 50       FKBP5 2.600000e-01 3.873333e+00    6.7433333 8.780000e+00 1.735000e+01
#> 51         FN1 7.500000e+00 0.000000e+00    0.0000000 0.000000e+00 0.000000e+00
#> 52       FOSL2 4.700000e-01 2.336667e+00    6.9333333 5.344000e+01 2.150633e+02
#> 53        GPD1 6.761333e+01 1.976267e+02  225.5300000 2.396900e+02 1.383833e+02
#> 54       GRINA 9.146667e+00 2.759000e+01   34.6400000 3.969667e+01 3.518333e+01
#> 55        HAMP 3.459333e+01 2.940033e+02  259.5933333 8.544833e+02 3.670600e+02
#> 56        HEPH 4.666667e-02 7.816667e+01  199.2533333 1.587333e+02 5.314000e+01
#> 57      HIGD1A 2.923333e+00 1.084667e+01   15.8400000 3.858000e+01 1.708033e+02
#> 58         HLF 1.540000e+00 5.933333e+00    7.1333333 7.926667e+00 3.703333e+00
#> 59       HMGCR 1.721667e+01 0.000000e+00    0.0000000 0.000000e+00 0.000000e+00
#> 60      HMGCS1 3.083333e+00 1.604000e+01   14.3500000 1.459667e+01 1.418667e+01
#> 61        HMOX 1.991100e+02 8.492667e+01   62.0300000 5.946333e+01 2.989667e+01
#> 62         HPX 0.000000e+00 3.217050e+03 7689.9733330 4.359253e+03 2.452460e+03
#> 63    HS3ST3B1 5.536667e+00 1.311000e+01   12.4566667 1.306333e+01 4.148667e+01
#> 64     HSD17B7 1.443333e+00 5.323333e+00    5.8700000 7.596667e+00 5.740000e+00
#> 65       HSP30 1.283333e+00 3.666667e-02    0.0000000 0.000000e+00 0.000000e+00
#> 66        IDI1 5.606667e+00 1.845333e+01   15.5133333 1.526667e+01 1.263333e+01
#> 67        IDO2 3.082667e+01 1.294000e+01   12.9566667 9.090000e+00 7.886667e+00
#> 68       IL1R1 1.765667e+01 3.945333e+01   76.0100000 6.639667e+01 8.238667e+01
#> 69        IL6R 1.513667e+01 3.547333e+01   42.8200000 6.374333e+01 9.317333e+01
#> 70      INSIG1 1.663333e+00 4.726667e+00   11.8066667 2.555333e+01 1.421000e+02
#> 71       ITIH3 1.773333e+01 0.000000e+00    0.0000000 0.000000e+00 0.000000e+00
#> 72       KLF12 2.383333e+00 8.160000e+00   10.7733333 1.068333e+01 1.653667e+01
#> 73       KRT18 5.518333e+01 5.666667e-02    0.0000000 1.333333e-02 2.666667e-02
#> 74      L2HGDH 8.000000e-01 2.178333e+01   57.4566667 5.222333e+01 5.741000e+01
#> 75       LECT2 8.176667e+00 2.443333e+01   27.0066667 6.048333e+01 7.552000e+01
#> 76        LRR1 4.073333e+00 1.256667e+00    0.7000000 1.020000e+00 3.300000e-01
#> 77       MAPK6 8.380000e+00 3.573333e+01   76.6933333 1.163033e+02 1.661600e+02
#> 78         MDH 6.616000e+01 1.887200e+02  283.3433333 2.673833e+02 3.673400e+02
#> 79    MID1IP1L 1.909000e+01 4.890000e+00    2.2366667 2.343333e+00 1.860000e+00
#> 80       MKNK2 6.093333e+00 1.235667e+01   23.9800000 3.382667e+01 4.504000e+01
#> 81     MT-ATP6 2.718863e+03 1.140000e+00    0.0000000 8.633333e-01 2.200000e+00
#> 82      MT-CO1 6.518067e+02 2.390333e+01   19.3933333 2.637000e+01 3.173000e+01
#> 83      MT-ND2 6.851533e+02 3.200000e-01    0.2533333 1.033333e-01 4.000000e-02
#> 84      MT-ND4 5.888367e+02 0.000000e+00    0.0000000 0.000000e+00 0.000000e+00
#> 85      MT-ND6 3.056333e+01 0.000000e+00    0.0000000 0.000000e+00 0.000000e+00
#> 86       NFIL3 1.169000e+01 2.749333e+01   30.9066667 6.079333e+01 5.792000e+01
#> 87        NOCT 4.153333e+00 1.972000e+01   23.4700000 5.542333e+01 5.076333e+01
#> 88        NQO1 2.864667e+01 1.249333e+01   13.7133333 1.344000e+01 9.870000e+00
#> 89       NUDT7 1.756667e+00 7.466667e-01    0.5300000 3.000000e-01 1.000000e-01
#> 90    P33MONOX 1.109667e+01 1.563333e+00    2.1633333 2.180000e+00 3.383333e+00
#> 91      PAIP2B 6.963333e+00 2.086667e+01   29.3500000 2.345333e+01 1.577667e+01
#> 92       PDIA3 9.583333e+00 2.512000e+01   19.7433333 2.295333e+01 4.403667e+01
#> 93        PDK2 5.918333e+01 1.528167e+02  124.2533333 1.802400e+02 1.472800e+02
#> 94      PDLIM2 1.085000e+01 4.200000e+00    3.1966667 2.483333e+00 5.700000e-01
#> 95        PIM3 4.753333e+00 1.441667e+01   36.9266667 7.831000e+01 8.694667e+01
#> 96       PLIN2 1.754000e+01 6.443000e+01  114.3400000 2.812067e+02 3.339400e+02
#> 97         PNP 7.322333e+01 3.324467e+02  881.4700000 1.394310e+03 1.597117e+03
#> 98       PPARA 2.170000e+00 1.288333e+01   17.8166667 2.622333e+01 1.314867e+02
#> 99      PPDPFA 1.083967e+02 2.487467e+02  387.0666667 3.986433e+02 2.438900e+02
#> 100      PPM1K 1.283333e+00 9.053333e+00   12.0200000 9.706667e+00 1.798333e+01
#> 101   PPP1CC-A 1.563333e+00 7.220000e+00   15.2600000 2.397333e+01 9.563333e+00
#> 102    PPP1R3B 3.904667e+01 1.653433e+02  221.6600000 3.285567e+02 1.404367e+02
#> 103      PPTC7 4.566667e-01 8.403333e+00    8.6800000 3.640000e+00 1.416667e+00
#> 104   PRELID3B 2.773333e+00 2.391333e+01   57.1700000 1.294000e+02 2.920267e+02
#> 105      PRODH 1.733333e-01 2.240000e+00   20.3933333 3.121667e+01 2.995167e+02
#> 106       PSPH 2.186667e+00 4.426667e+00   10.0933333 9.340000e+00 2.203333e+01
#> 107      RASD2 2.800000e-01 5.140000e+00    3.8000000 1.191333e+01 1.950333e+01
#> 108       RBMX 1.384333e+01 6.116667e+00    3.8933333 3.146667e+00 3.310000e+00
#> 109      RBP4A 2.902333e+02 1.231233e+02   96.4600000 1.098633e+02 3.456333e+01
#> 110     RHBG-A 2.213333e+01 8.630000e+00    3.7033333 4.736667e+00 1.863333e+00
#> 111      RHOAD 3.616667e+00 9.700000e+00    8.9833333 7.886667e+00 8.776667e+00
#> 112       RND3 8.746667e+00 2.095333e+01   28.2900000 3.600000e+01 5.347667e+01
#> 113     RNF223 1.900000e-01 2.953333e+00    2.1033333 2.233333e+00 4.996667e+00
#> 114      RRBP1 1.441667e+01 2.974667e+01   41.7333333 3.487667e+01 6.048333e+01
#> 115      RXRGB 3.623667e+01 1.136333e+01   13.1933333 1.239333e+01 6.796667e+00
#> 116       SAT1 1.087333e+01 2.813667e+01   67.0366667 1.804133e+02 2.885800e+02
#> 117 SERPINA3-3 0.000000e+00 1.429667e+01   55.0900000 1.235400e+02 7.930000e+00
#> 118 SERPINA3-8 3.450667e+01 6.908667e+01  110.0633333 1.765733e+02 3.544933e+02
#> 119   SERPING1 0.000000e+00 2.666667e+01   61.7333333 3.017333e+01 5.588000e+01
#> 120       SFPQ 1.734333e+01 3.323667e+01   52.9933333 6.145000e+01 8.585333e+01
#> 121    SLC16A1 4.700000e-01 1.210333e+01   23.8233333 3.028333e+01 8.282000e+01
#> 122   SLC20A1B 6.540000e+00 1.882333e+01   35.7666667 5.276333e+01 2.136067e+02
#> 123    SLC22A2 2.548000e+01 7.450000e+00    5.4633333 5.120000e+00 4.363333e+00
#> 124  SLC25A25A 9.500000e-01 2.232000e+01   64.2633333 1.125200e+02 1.148300e+02
#> 125    SLC2A12 1.145000e+01 5.053333e+00    4.2700000 3.610000e+00 7.300000e-01
#> 126     SLC3A2 3.533333e+00 7.610000e+00   13.1033333 1.872667e+01 6.079333e+01
#> 127      SOAT1 1.577000e+01 6.882667e+01  129.1366667 8.004333e+01 4.728000e+01
#> 128      SOCS3 1.663333e+00 4.130000e+00    4.5833333 1.426333e+01 1.410000e+01
#> 129       SSR1 0.000000e+00 1.529333e+01   22.6800000 1.376000e+01 1.796667e+01
#> 130        TAT 5.766667e-01 6.603333e+00    9.8166667 1.084000e+01 7.493333e+00
#> 131      TCF21 6.530000e+00 2.966667e+00    1.8533333 1.273333e+00 1.393333e+00
#> 132         TF 0.000000e+00 1.085793e+03  954.0600000 2.406800e+03 5.783067e+02
#> 133      TOR3A 1.106667e+00 8.290000e+00    7.3400000 1.264333e+01 3.005333e+01
#> 134       UGT3 1.873333e+00 6.533333e+00   13.6433333 6.513333e+00 1.601000e+01
#> 135       UPP2 3.858333e+01 1.067800e+02  146.2866667 1.365233e+02 1.321167e+02
#> 136       UROS 7.863333e+00 2.613333e+00    1.6300000 7.866667e-01 6.200000e-01
#> 137       XBP1 6.284333e+01 1.555433e+02  164.5566667 1.905633e+02 2.309567e+02
#> 138     YME1L1 9.370000e+00 1.831000e+01   29.1133333 3.102000e+01 5.596000e+01
#> 139      YPEL3 2.465333e+01 6.273333e+00    4.9033333 4.913333e+00 1.080000e+00
#>                   Pathways
#> 1   PPAR signaling pathway
#> 2   PPAR signaling pathway
#> 3   PPAR signaling pathway
#> 4   PPAR signaling pathway
#> 5   PPAR signaling pathway
#> 6   PPAR signaling pathway
#> 7   PPAR signaling pathway
#> 8   PPAR signaling pathway
#> 9   PPAR signaling pathway
#> 10  PPAR signaling pathway
#> 11  PPAR signaling pathway
#> 12  PPAR signaling pathway
#> 13  PPAR signaling pathway
#> 14  PPAR signaling pathway
#> 15  PPAR signaling pathway
#> 16  PPAR signaling pathway
#> 17  PPAR signaling pathway
#> 18  PPAR signaling pathway
#> 19  PPAR signaling pathway
#> 20  PPAR signaling pathway
#> 21  PPAR signaling pathway
#> 22  PPAR signaling pathway
#> 23  PPAR signaling pathway
#> 24  PPAR signaling pathway
#> 25  PPAR signaling pathway
#> 26  PPAR signaling pathway
#> 27  PPAR signaling pathway
#> 28  PPAR signaling pathway
#> 29  PPAR signaling pathway
#> 30  PPAR signaling pathway
#> 31  PPAR signaling pathway
#> 32  PPAR signaling pathway
#> 33  PPAR signaling pathway
#> 34  PPAR signaling pathway
#> 35  PPAR signaling pathway
#> 36  PPAR signaling pathway
#> 37  PPAR signaling pathway
#> 38  PPAR signaling pathway
#> 39  PPAR signaling pathway
#> 40  PPAR signaling pathway
#> 41  PPAR signaling pathway
#> 42  PPAR signaling pathway
#> 43  PPAR signaling pathway
#> 44  PPAR signaling pathway
#> 45  PPAR signaling pathway
#> 46  PPAR signaling pathway
#> 47  PPAR signaling pathway
#> 48  PPAR signaling pathway
#> 49  PPAR signaling pathway
#> 50  PPAR signaling pathway
#> 51            AMPK pathway
#> 52            AMPK pathway
#> 53            AMPK pathway
#> 54            AMPK pathway
#> 55            AMPK pathway
#> 56            AMPK pathway
#> 57            AMPK pathway
#> 58            AMPK pathway
#> 59            AMPK pathway
#> 60            AMPK pathway
#> 61            AMPK pathway
#> 62            AMPK pathway
#> 63            AMPK pathway
#> 64            AMPK pathway
#> 65            AMPK pathway
#> 66            AMPK pathway
#> 67            AMPK pathway
#> 68            AMPK pathway
#> 69            AMPK pathway
#> 70            AMPK pathway
#> 71            AMPK pathway
#> 72            AMPK pathway
#> 73            AMPK pathway
#> 74            AMPK pathway
#> 75            AMPK pathway
#> 76            AMPK pathway
#> 77            AMPK pathway
#> 78            AMPK pathway
#> 79            AMPK pathway
#> 80            AMPK pathway
#> 81            AMPK pathway
#> 82            AMPK pathway
#> 83            AMPK pathway
#> 84            AMPK pathway
#> 85            AMPK pathway
#> 86            AMPK pathway
#> 87            AMPK pathway
#> 88            AMPK pathway
#> 89            AMPK pathway
#> 90            AMPK pathway
#> 91            AMPK pathway
#> 92            AMPK pathway
#> 93            AMPK pathway
#> 94            AMPK pathway
#> 95            AMPK pathway
#> 96            AMPK pathway
#> 97            AMPK pathway
#> 98            AMPK pathway
#> 99            AMPK pathway
#> 100           AMPK pathway
#> 101      Autophagy pathway
#> 102      Autophagy pathway
#> 103      Autophagy pathway
#> 104      Autophagy pathway
#> 105      Autophagy pathway
#> 106      Autophagy pathway
#> 107      Autophagy pathway
#> 108      Autophagy pathway
#> 109      Autophagy pathway
#> 110      Autophagy pathway
#> 111      Autophagy pathway
#> 112      Autophagy pathway
#> 113      Autophagy pathway
#> 114      Autophagy pathway
#> 115      Autophagy pathway
#> 116      Autophagy pathway
#> 117      Autophagy pathway
#> 118      Autophagy pathway
#> 119      Autophagy pathway
#> 120      Autophagy pathway
#> 121      Autophagy pathway
#> 122      Autophagy pathway
#> 123      Autophagy pathway
#> 124      Autophagy pathway
#> 125      Autophagy pathway
#> 126      Autophagy pathway
#> 127      Autophagy pathway
#> 128      Autophagy pathway
#> 129      Autophagy pathway
#> 130      Autophagy pathway
#> 131      Autophagy pathway
#> 132      Autophagy pathway
#> 133      Autophagy pathway
#> 134      Autophagy pathway
#> 135      Autophagy pathway
#> 136      Autophagy pathway
#> 137      Autophagy pathway
#> 138      Autophagy pathway
#> 139      Autophagy pathway