Table merge used to merge multiple variables to on variable.

Table merge used to merge multiple variables to on variable.

Usage

table_merge(
  data,
  merge_vars = c("Ozone", "Solar.R", "Wind", "Temp"),
  new_var = "Variable",
  new_value = "Value",
  na_remove = FALSE
)

Arguments

data

Dataframe: include multiple variables (columns).

merge_vars

Vector: include merged variable (column) names. Default: c("Ozone", "Solar.R", "Wind", "Temp") in example data.

new_var

Character: new variable (column) name. Default: "Variable".

new_value

Character: new variable (column) value name. Default: "Value".

na_remove

Logical: remove NA value. Default: FALSE, options: TRUE, FALSE.

Value

Table: include multiple variables.

Author

benben-miao

Examples

# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Use example dataset
data(table_merge_data)
head(table_merge_data)
#>   Ozone Solar.R Wind Temp Month Day
#> 1    41     190  7.4   67     5   1
#> 2    36     118  8.0   72     5   2
#> 3    12     149 12.6   74     5   3
#> 4    18     313 11.5   62     5   4
#> 5    NA      NA 14.3   56     5   5
#> 6    28      NA 14.9   66     5   6

# 3. Default parameters
res <- table_merge(table_merge_data)
head(res)
#>   Month Day Variable Value
#> 1     5   1    Ozone    41
#> 2     5   2    Ozone    36
#> 3     5   3    Ozone    12
#> 4     5   4    Ozone    18
#> 5     5   5    Ozone    NA
#> 6     5   6    Ozone    28

# 4. Set merge_vars = c("Ozone", "Solar.R")
res <- table_merge(table_merge_data, merge_vars = c("Ozone", "Solar.R"))
head(res)
#>   Wind Temp Month Day Variable Value
#> 1  7.4   67     5   1    Ozone    41
#> 2  8.0   72     5   2    Ozone    36
#> 3 12.6   74     5   3    Ozone    12
#> 4 11.5   62     5   4    Ozone    18
#> 5 14.3   56     5   5    Ozone    NA
#> 6 14.9   66     5   6    Ozone    28