Trend data include genes expression in multiple traits as example for trend_plot function.

Trend data include genes expression in multiple traits as example for trend_plot function.

Usage

data(trend_data)

Format

Dataframe: include columns (Gene, Trait1, Trait2, ..., Pathway), rows (Genes).

References

https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/ExpressionTrend/

Author

benben-miao

Examples

# 1. Library TOmicsVis package
library(TOmicsVis)

# 2. Load example data
data(trend_data)

# 3. View example data
trend_data
#>        Name Trait1 Trait2 Trait3                Pathway
#> 1     gene1    5.1    1.4    3.5 PPAR signaling pathway
#> 2     gene2    4.9    1.4    3.0 PPAR signaling pathway
#> 3     gene3    4.7    1.3    3.2 PPAR signaling pathway
#> 4     gene4    4.6    1.5    3.1 PPAR signaling pathway
#> 5     gene5    5.0    1.4    3.6 PPAR signaling pathway
#> 6     gene6    5.4    1.7    3.9 PPAR signaling pathway
#> 7     gene7    4.6    1.4    3.4 PPAR signaling pathway
#> 8     gene8    5.0    1.5    3.4 PPAR signaling pathway
#> 9     gene9    4.4    1.4    2.9 PPAR signaling pathway
#> 10   gene10    4.9    1.5    3.1 PPAR signaling pathway
#> 11   gene11    5.4    1.5    3.7 PPAR signaling pathway
#> 12   gene12    4.8    1.6    3.4 PPAR signaling pathway
#> 13   gene13    4.8    1.4    3.0 PPAR signaling pathway
#> 14   gene14    4.3    1.1    3.0 PPAR signaling pathway
#> 15   gene15    5.8    1.2    4.0 PPAR signaling pathway
#> 16   gene16    5.7    1.5    4.4 PPAR signaling pathway
#> 17   gene17    5.4    1.3    3.9 PPAR signaling pathway
#> 18   gene18    5.1    1.4    3.5 PPAR signaling pathway
#> 19   gene19    5.7    1.7    3.8 PPAR signaling pathway
#> 20   gene20    5.1    1.5    3.8 PPAR signaling pathway
#> 21   gene21    5.4    1.7    3.4 PPAR signaling pathway
#> 22   gene22    5.1    1.5    3.7 PPAR signaling pathway
#> 23   gene23    4.6    1.0    3.6 PPAR signaling pathway
#> 24   gene24    5.1    1.7    3.3 PPAR signaling pathway
#> 25   gene25    4.8    1.9    3.4 PPAR signaling pathway
#> 26   gene26    5.0    1.6    3.0 PPAR signaling pathway
#> 27   gene27    5.0    1.6    3.4 PPAR signaling pathway
#> 28   gene28    5.2    1.5    3.5 PPAR signaling pathway
#> 29   gene29    5.2    1.4    3.4 PPAR signaling pathway
#> 30   gene30    4.7    1.6    3.2 PPAR signaling pathway
#> 31   gene31    4.8    1.6    3.1 PPAR signaling pathway
#> 32   gene32    5.4    1.5    3.4 PPAR signaling pathway
#> 33   gene33    5.2    1.5    4.1 PPAR signaling pathway
#> 34   gene34    5.5    1.4    4.2 PPAR signaling pathway
#> 35   gene35    4.9    1.5    3.1 PPAR signaling pathway
#> 36   gene36    5.0    1.2    3.2 PPAR signaling pathway
#> 37   gene37    5.5    1.3    3.5 PPAR signaling pathway
#> 38   gene38    4.9    1.4    3.6 PPAR signaling pathway
#> 39   gene39    4.4    1.3    3.0 PPAR signaling pathway
#> 40   gene40    5.1    1.5    3.4 PPAR signaling pathway
#> 41   gene41    5.0    1.3    3.5 PPAR signaling pathway
#> 42   gene42    4.5    1.3    2.3 PPAR signaling pathway
#> 43   gene43    4.4    1.3    3.2 PPAR signaling pathway
#> 44   gene44    5.0    1.6    3.5 PPAR signaling pathway
#> 45   gene45    5.1    1.9    3.8 PPAR signaling pathway
#> 46   gene46    4.8    1.4    3.0 PPAR signaling pathway
#> 47   gene47    5.1    1.6    3.8 PPAR signaling pathway
#> 48   gene48    4.6    1.4    3.2 PPAR signaling pathway
#> 49   gene49    5.3    1.5    3.7 PPAR signaling pathway
#> 50   gene50    5.0    1.4    3.3 PPAR signaling pathway
#> 51   gene51    7.0    4.7    3.2           AMPK pathway
#> 52   gene52    6.4    4.5    3.2           AMPK pathway
#> 53   gene53    6.9    4.9    3.1           AMPK pathway
#> 54   gene54    5.5    4.0    2.3           AMPK pathway
#> 55   gene55    6.5    4.6    2.8           AMPK pathway
#> 56   gene56    5.7    4.5    2.8           AMPK pathway
#> 57   gene57    6.3    4.7    3.3           AMPK pathway
#> 58   gene58    4.9    3.3    2.4           AMPK pathway
#> 59   gene59    6.6    4.6    2.9           AMPK pathway
#> 60   gene60    5.2    3.9    2.7           AMPK pathway
#> 61   gene61    5.0    3.5    2.0           AMPK pathway
#> 62   gene62    5.9    4.2    3.0           AMPK pathway
#> 63   gene63    6.0    4.0    2.2           AMPK pathway
#> 64   gene64    6.1    4.7    2.9           AMPK pathway
#> 65   gene65    5.6    3.6    2.9           AMPK pathway
#> 66   gene66    6.7    4.4    3.1           AMPK pathway
#> 67   gene67    5.6    4.5    3.0           AMPK pathway
#> 68   gene68    5.8    4.1    2.7           AMPK pathway
#> 69   gene69    6.2    4.5    2.2           AMPK pathway
#> 70   gene70    5.6    3.9    2.5           AMPK pathway
#> 71   gene71    5.9    4.8    3.2           AMPK pathway
#> 72   gene72    6.1    4.0    2.8           AMPK pathway
#> 73   gene73    6.3    4.9    2.5           AMPK pathway
#> 74   gene74    6.1    4.7    2.8           AMPK pathway
#> 75   gene75    6.4    4.3    2.9           AMPK pathway
#> 76   gene76    6.6    4.4    3.0           AMPK pathway
#> 77   gene77    6.8    4.8    2.8           AMPK pathway
#> 78   gene78    6.7    5.0    3.0           AMPK pathway
#> 79   gene79    6.0    4.5    2.9           AMPK pathway
#> 80   gene80    5.7    3.5    2.6           AMPK pathway
#> 81   gene81    5.5    3.8    2.4           AMPK pathway
#> 82   gene82    5.5    3.7    2.4           AMPK pathway
#> 83   gene83    5.8    3.9    2.7           AMPK pathway
#> 84   gene84    6.0    5.1    2.7           AMPK pathway
#> 85   gene85    5.4    4.5    3.0           AMPK pathway
#> 86   gene86    6.0    4.5    3.4           AMPK pathway
#> 87   gene87    6.7    4.7    3.1           AMPK pathway
#> 88   gene88    6.3    4.4    2.3           AMPK pathway
#> 89   gene89    5.6    4.1    3.0           AMPK pathway
#> 90   gene90    5.5    4.0    2.5           AMPK pathway
#> 91   gene91    5.5    4.4    2.6           AMPK pathway
#> 92   gene92    6.1    4.6    3.0           AMPK pathway
#> 93   gene93    5.8    4.0    2.6           AMPK pathway
#> 94   gene94    5.0    3.3    2.3           AMPK pathway
#> 95   gene95    5.6    4.2    2.7           AMPK pathway
#> 96   gene96    5.7    4.2    3.0           AMPK pathway
#> 97   gene97    5.7    4.2    2.9           AMPK pathway
#> 98   gene98    6.2    4.3    2.9           AMPK pathway
#> 99   gene99    5.1    3.0    2.5           AMPK pathway
#> 100 gene100    5.7    4.1    2.8           AMPK pathway
#> 101 gene101    6.3    6.0    3.3      Autophagy pathway
#> 102 gene102    5.8    5.1    2.7      Autophagy pathway
#> 103 gene103    7.1    5.9    3.0      Autophagy pathway
#> 104 gene104    6.3    5.6    2.9      Autophagy pathway
#> 105 gene105    6.5    5.8    3.0      Autophagy pathway
#> 106 gene106    7.6    6.6    3.0      Autophagy pathway
#> 107 gene107    4.9    4.5    2.5      Autophagy pathway
#> 108 gene108    7.3    6.3    2.9      Autophagy pathway
#> 109 gene109    6.7    5.8    2.5      Autophagy pathway
#> 110 gene110    7.2    6.1    3.6      Autophagy pathway
#> 111 gene111    6.5    5.1    3.2      Autophagy pathway
#> 112 gene112    6.4    5.3    2.7      Autophagy pathway
#> 113 gene113    6.8    5.5    3.0      Autophagy pathway
#> 114 gene114    5.7    5.0    2.5      Autophagy pathway
#> 115 gene115    5.8    5.1    2.8      Autophagy pathway
#> 116 gene116    6.4    5.3    3.2      Autophagy pathway
#> 117 gene117    6.5    5.5    3.0      Autophagy pathway
#> 118 gene118    7.7    6.7    3.8      Autophagy pathway
#> 119 gene119    7.7    6.9    2.6      Autophagy pathway
#> 120 gene120    6.0    5.0    2.2      Autophagy pathway
#> 121 gene121    6.9    5.7    3.2      Autophagy pathway
#> 122 gene122    5.6    4.9    2.8      Autophagy pathway
#> 123 gene123    7.7    6.7    2.8      Autophagy pathway
#> 124 gene124    6.3    4.9    2.7      Autophagy pathway
#> 125 gene125    6.7    5.7    3.3      Autophagy pathway
#> 126 gene126    7.2    6.0    3.2      Autophagy pathway
#> 127 gene127    6.2    4.8    2.8      Autophagy pathway
#> 128 gene128    6.1    4.9    3.0      Autophagy pathway
#> 129 gene129    6.4    5.6    2.8      Autophagy pathway
#> 130 gene130    7.2    5.8    3.0      Autophagy pathway
#> 131 gene131    7.4    6.1    2.8      Autophagy pathway
#> 132 gene132    7.9    6.4    3.8      Autophagy pathway
#> 133 gene133    6.4    5.6    2.8      Autophagy pathway
#> 134 gene134    6.3    5.1    2.8      Autophagy pathway
#> 135 gene135    6.1    5.6    2.6      Autophagy pathway
#> 136 gene136    7.7    6.1    3.0      Autophagy pathway
#> 137 gene137    6.3    5.6    3.4      Autophagy pathway
#> 138 gene138    6.4    5.5    3.1      Autophagy pathway
#> 139 gene139    6.0    4.8    3.0      Autophagy pathway
#> 140 gene140    6.9    5.4    3.1      Autophagy pathway
#> 141 gene141    6.7    5.6    3.1      Autophagy pathway
#> 142 gene142    6.9    5.1    3.1      Autophagy pathway
#> 143 gene143    5.8    5.1    2.7      Autophagy pathway
#> 144 gene144    6.8    5.9    3.2      Autophagy pathway
#> 145 gene145    6.7    5.7    3.3      Autophagy pathway
#> 146 gene146    6.7    5.2    3.0      Autophagy pathway
#> 147 gene147    6.3    5.0    2.5      Autophagy pathway
#> 148 gene148    6.5    5.2    3.0      Autophagy pathway
#> 149 gene149    6.2    5.4    3.4      Autophagy pathway
#> 150 gene150    5.9    5.1    3.0      Autophagy pathway
#> 151 gene151    5.7    4.1    2.8           MAPK pathway
#> 152 gene152    6.3    6.0    3.3           MAPK pathway
#> 153 gene153    5.8    5.1    2.7           MAPK pathway
#> 154 gene154    7.1    5.9    3.0           MAPK pathway
#> 155 gene155    6.3    5.6    2.9           MAPK pathway
#> 156 gene156    6.5    5.8    3.0           MAPK pathway
#> 157 gene157    7.6    6.6    3.0           MAPK pathway
#> 158 gene158    4.9    4.5    2.5           MAPK pathway
#> 159 gene159    7.3    6.3    2.9           MAPK pathway
#> 160 gene160    6.7    5.8    2.5           MAPK pathway
#> 161 gene161    7.2    6.1    3.6           MAPK pathway
#> 162 gene162    6.5    5.1    3.2           MAPK pathway
#> 163 gene163    6.4    5.3    2.7           MAPK pathway
#> 164 gene164    6.8    5.5    3.0           MAPK pathway
#> 165 gene165    5.7    5.0    2.5           MAPK pathway
#> 166 gene166    5.8    5.1    2.8           MAPK pathway
#> 167 gene167    6.4    5.3    3.2           MAPK pathway
#> 168 gene168    6.5    5.5    3.0           MAPK pathway
#> 169 gene169    7.7    6.7    3.8           MAPK pathway
#> 170 gene170    7.7    6.9    2.6           MAPK pathway
#> 171 gene171    6.0    5.0    2.2           MAPK pathway
#> 172 gene172    6.9    5.7    3.2           MAPK pathway
#> 173 gene173    5.6    4.9    2.8           MAPK pathway
#> 174 gene174    7.7    6.7    2.8           MAPK pathway
#> 175 gene175    6.3    4.9    2.7           MAPK pathway
#> 176 gene176    6.7    5.7    3.3           MAPK pathway
#> 177 gene177    7.2    6.0    3.2           MAPK pathway
#> 178 gene178    6.2    4.8    2.8           MAPK pathway
#> 179 gene179    6.1    4.9    3.0           MAPK pathway
#> 180 gene180    6.4    5.6    2.8           MAPK pathway
#> 181 gene181    7.2    5.8    3.0           MAPK pathway
#> 182 gene182    7.4    6.1    2.8           MAPK pathway
#> 183 gene183    7.9    6.4    3.8           MAPK pathway
#> 184 gene184    6.4    5.6    2.8           MAPK pathway
#> 185 gene185    6.3    5.1    2.8           MAPK pathway
#> 186 gene186    6.1    5.6    2.6           MAPK pathway
#> 187 gene187    7.7    6.1    3.0           MAPK pathway
#> 188 gene188    6.3    5.6    3.4           MAPK pathway
#> 189 gene189    6.4    5.5    3.1           MAPK pathway
#> 190 gene190    6.0    4.8    3.0           MAPK pathway
#> 191 gene191    6.9    5.4    3.1           MAPK pathway
#> 192 gene192    6.7    5.6    3.1           MAPK pathway
#> 193 gene193    6.9    5.1    3.1           MAPK pathway
#> 194 gene194    5.8    5.1    2.7           MAPK pathway
#> 195 gene195    6.8    5.9    3.2           MAPK pathway
#> 196 gene196    6.7    5.7    3.3           MAPK pathway
#> 197 gene197    6.7    5.2    3.0           MAPK pathway
#> 198 gene198    6.3    5.0    2.5           MAPK pathway
#> 199 gene199    6.5    5.2    3.0           MAPK pathway
#> 200 gene200    6.2    5.4    3.4           MAPK pathway